+++ app/routes/account/benefits/carrier-link.js
@@ -1,5 +1,5 @@
 import Route from 'granite/core/route';
 
 export default class CarrierLinkRoute extends Route {
- titleToken='Carriers'
+ titleToken = 'Carriers'
 }
