+++ app/templates/account/benefits.hbs
@@ -1,9 +1,9 @@
 <div class="ui container basic segment">
   <UiMenu @menuClass="pointing menu__benefits-page">
-    <LinkTo @route="account.benefits" class="item">
+    <LinkTo @route="account.benefits.index" class="item">
       <i class="list icon"></i>Summary
     </LinkTo>
-    <LinkTo @route="account.benefits.plans.integrations" class="item">
+    <LinkTo @route="account.benefits.plans" class="item">
       <i class="cubes icon"></i>Plans
     </LinkTo>
     <LinkTo @route="account.benefits.enrollments" class="item">
