--- app/templates/account/benefits/carrier-link.hbs
@@ -1,1 +1,0 @@
-{{outlet}}
